refactor: split api_server.py into routers + add agent.run() unit tests
api_server.py (1911 lines) → thin entry point (290 lines):
- evolving_agent/utils/schemas.py  — all 20 Pydantic models
- evolving_agent/utils/app_state.py — shared mutable globals (agent, github_modifier, etc.)
- evolving_agent/utils/deps.py     — get_agent, verify_api_key dependencies
- evolving_agent/api/routes/       — 9 router modules by domain:
  general, interaction, memory, knowledge, self_improvement,
  github, discord, web_search, system

Backward compat preserved: `from evolving_agent.utils.api_server import app, get_agent`
and `api_server_module.agent` still work.

tests/test_agent_run.py: 16 unit tests for SelfImprovingAgent.run() covering:
- uninitialized guard, happy path, interaction count, context retrieval,
  conversation history fetching, evaluation on/off, self-edit detection,
  self-modification trigger cadence, exception propagation, knowledge update
